# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "2gis-calm-map API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/organization": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get organization for current authenticated owner (admin/owner). 1:1 per owner.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Get organization (owner self)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Organization"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create organization for current user (1:1). Roles: owner, admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Create organization",
                "parameters": [
                    {
                        "description": "Organization data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.OrganizationCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Organization"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially update organization (owner/admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Update organization",
                "parameters": [
                    {
                        "description": "Fields to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.OrganizationUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Organization"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/organization/params/average/by-type": {
            "post": {
                "description": "For every organization of a specified type computes (avg(p1)+...)/N and returns only those with average \u003e threshold (default 3.0).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization-params"
                ],
                "summary": "Compute averages for each organization of given type",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.OrganizationsParamsAverageByTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.OrganizationsParamsAverageByTypeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/organization/params/average/with-info": {
            "post": {
                "description": "Как /organization/params/average, но дополняет адресом/типом/координатами и путями изображений. Публично.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization-params"
                ],
                "summary": "Compute average and return organization info",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.OrganizationParamsWithOrgRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.OrganizationParamsWithOrgResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/organization/public/by-address": {
            "post": {
                "description": "Возвращает организацию (id и основные поля) по точному адресу. Публично.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Public organization lookup by address",
                "parameters": [
                    {
                        "description": "Address lookup",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.OrganizationByAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.OrganizationByAddressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.OrganizationByAddressRequest": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
        "handler.OrganizationByAddressResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "map_path": {
                    "type": "string"
                },
                "organization_type": {
                    "type": "string"
                },
                "picture_path": {
                    "type": "string"
                }
            }
        },
...
        "handler.OrganizationCreateRequest": {
            "type": "object",
            "required": [
                "address",
                "organization_type"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "organization_type": {
                    "type": "string"
                }
            }
        },
...
        "handler.OrganizationParamsWithOrgRequest": {
            "type": "object",
            "required": [
                "organization_id",
                "params"
            ],
            "properties": {
                "organization_id": {
                    "type": "integer"
                },
                "params": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handler.OrganizationParamsWithOrgResponse": {
            "type": "object",
            "properties": {
                "average": {
                    "type": "number"
                },
                "organization": {
                    "type": "object",
                    "properties": {
                        "address": {
                            "type": "string"
                        },
                        "id": {
                            "type": "integer"
                        },
                        "latitude": {
                            "type": "number"
                        },
                        "longitude": {
                            "type": "number"
                        },
                        "map_path": {
                            "type": "string"
                        },
                        "organization_type": {
                            "type": "string"
                        },
                        "picture_path": {
                            "type": "string"
                        }
                    }
                },
                "params": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handler.OrganizationUpdateRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "organization_type": {
                    "type": "string"
                }
            }
        },
        "handler.OrganizationWithSelectedAverage": {
            "type": "object",
            "properties": {
                "average": {
                    "type": "number"
                },
                "organization": {},
                "params": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handler.OrganizationsParamsAverageByTypeRequest": {
            "type": "object",
            "required": [
                "organization_type",
                "params"
            ],
            "properties": {
                "organization_type": {
                    "type": "string"
                },
                "params": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "threshold": {
                    "description": "Optional threshold; if omitted, defaults to 3.0",
                    "type": "number"
                }
            }
        },
        "handler.OrganizationsParamsAverageByTypeResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.OrganizationWithSelectedAverage"
                    }
                },
                "organization_type": {
                    "type": "string"
                }
            }
        },
...
        "model.Organization": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "description": "optional",
                    "type": "number"
                },
                "longitude": {
                    "description": "optional",
                    "type": "number"
                },
                "map_path": {
                    "description": "относительный путь к карте (изображение)",
                    "type": "string"
                },
                "organization_type": {
                    "type": "string"
                },
                "owner_id": {
                    "description": "один владелец - одна организация",
                    "type": "integer"
                },
                "params": {
                    "$ref": "#/definitions/model.OrganizationParams"
                },
                "picture_path": {
                    "description": "относительный путь к общей картинке",
                    "type": "string"
                }
            }
        },
...
        "model.OrganizationParams": {
            "type": "object",
            "properties": {
                "appearance_avg": {
                    "type": "number"
                },
                "appearance_count": {
                    "type": "integer"
                },
                "appearance_sum": {
                    "type": "integer"
                },
                "calmness_avg": {
                    "type": "number"
                },
                "calmness_count": {
                    "type": "integer"
                },
                "calmness_sum": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "intuitiveness_avg": {
                    "type": "number"
                },
                "intuitiveness_count": {
                    "type": "integer"
                },
                "intuitiveness_sum": {
                    "type": "integer"
                },
                "lighting_avg": {
                    "type": "number"
                },
                "lighting_count": {
                    "type": "integer"
                },
                "lighting_sum": {
                    "type": "integer"
                },
                "organization_id": {
                    "type": "integer"
                },
                "people_density_avg": {
                    "type": "number"
                },
                "people_density_count": {
                    "type": "integer"
                },
                "people_density_sum": {
                    "type": "integer"
                },
                "self_service_avg": {
                    "type": "number"
                },
                "self_service_count": {
                    "type": "integer"
                },
                "self_service_sum": {
                    "type": "integer"
                },
                "signage_avg": {
                    "type": "number"
                },
                "signage_count": {
                    "type": "integer"
                },
                "signage_sum": {
                    "type": "integer"
                },
                "smell_avg": {
                    "type": "number"
                },
                "smell_count": {
                    "type": "integer"
                },
                "smell_sum": {
                    "type": "integer"
                },
                "staff_attitude_avg": {
                    "type": "number"
                },
                "staff_attitude_count": {
                    "type": "integer"
                },
                "staff_attitude_sum": {
                    "type": "integer"
                },
                "tactility_avg": {
                    "type": "number"
                },
                "tactility_count": {
                    "type": "integer"
                },
                "tactility_sum": {
                    "type": "integer"
                },
                "temperature_avg": {
                    "type": "number"
                },
                "temperature_count": {
                    "type": "integer"
                },
                "temperature_sum": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}